move the Titlte compenent to the Component folder
