+++ UI/ui/src/Notes/Notes.js
@@ -1,7 +1,7 @@
 import './Notes.css';
 import React,{ use, useState } from 'react';
 import Calendar from 'react-calendar'
-import PageTitle from '../PageTitle/PageTitle';
+import PageTitle from '../Components/PageTitle/PageTitle';
 import 'react-calendar/dist/Calendar.css';
 import axios from 'axios';
 import Popup from 'reactjs-popup';
