+++ UI/ui/src/Profile/Profile.js
@@ -1,6 +1,6 @@
 import './Profile.css';
 import React from 'react';
-
+import PageTitle from '../Components/PageTitle/PageTitle';
 const Profile = () => {
     return(
        <p>Notes</p>
