+++ UI/ui/src/Statistics/Statistics.js
@@ -1,6 +1,6 @@
 import './Statistics.css';
 import React from 'react';
-
+import PageTitle from '../Components/PageTitle/PageTitle';
 const Statistics = () => {
     return(
        <p>Notes</p>
